+++ CHANGELOG.md
@@ -4,11 +4,20 @@
 
 ### Features
 
+- Able to change/edit account name in accountDetails page
+
 ### Fixes
+
+- change staking and spending token ids
 - Fix switch local network some time not working issue
 
 ### Improvements
 
+# v0.9.1 (2019.04.08)
+
+### Features
+
+- upgrade cennznet-node to 0.9.17
 
 # v0.9.0 (2019.04.05)